bundle in house functions to one util script
